Generate board via a hook
